% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractaltree-package.R
\docType{package}
\name{fractaltree-package}
\alias{fractaltree}
\alias{fractaltree-package}
\title{fractaltree: Creates fractal trees for plotting}
\description{
Visualize neural net architectures using the "ggraph" package. 
    Creates and plots a graph using the layer and node information.  Right now
    only supports densely connected layers.
}
\author{
\strong{Maintainer}: Andrie de Vries \email{apdevries@gmail.com}

}
\keyword{internal}
