% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow.R
\name{grow}
\alias{grow}
\title{Grow a branch}
\usage{
grow(m, leaf, growth_fraction = 1)
}
\arguments{
\item{m}{input matrix of}

\item{leaf}{leaf shape, a matrix with 4 columns for x, y, xend, yend}

\item{growth_fraction}{}
}
\description{
Grow a branch
}
