% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{kaleidoscope}
\alias{kaleidoscope}
\title{Translate object in x and y plane}
\usage{
kaleidoscope(m, n)
}
\arguments{
\item{m}{Input matrix}

\item{xy}{a Vector of length two with x and y component}
}
\description{
Translate object in x and y plane
}
\seealso{
Other Rotate, translate and shrink: \code{\link{rotate}},
  \code{\link{shrink}}, \code{\link{translate}}
}
\concept{Rotate, translate and shrink}
