% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow.R
\name{plot_tree}
\alias{plot_tree}
\title{Plot the tree}
\usage{
plot_tree(x, colors = c("brown", "darkgreen"), colours = colors)
}
\arguments{
\item{x}{A tree object, resulting from \code{grow_tree()}}

\item{colors}{A vector of lenght two with start and end colour}

\item{colours}{}
}
\description{
Plot the tree
}
