% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{rotate}
\alias{rotate}
\title{Rotate object using matrix algebra.}
\usage{
rotate(m, theta = 0, radians = FALSE)
}
\arguments{
\item{m}{Input matrix}

\item{theta}{Angle of rotation}

\item{radians}{If TRUE theta is radians, otherwise degrees}
}
\description{
Rotate object using matrix algebra.
}
\seealso{
Other Rotate, translate and shrink: \code{\link{shrink}},
  \code{\link{translate}}
}
\concept{Rotate, translate and shrink}
