% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{shrink}
\alias{shrink}
\title{Shrink object}
\usage{
shrink(m, s = 0)
}
\arguments{
\item{m}{Input matrix}

\item{s}{scaling factor}
}
\value{
foo
}
\description{
Shrink object
}
\seealso{
Other Rotate, translate and shrink: 
\code{\link{kaleidoscope}()},
\code{\link{rotate}()},
\code{\link{translate}()}
}
\concept{Rotate, translate and shrink}
