% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{translate}
\alias{translate}
\title{Translate object}
\usage{
translate(m, xy)
}
\arguments{
\item{m}{Input matrix}

\item{s}{scaling factor}
}
\description{
Translate object
}
\seealso{
Other Rotate, translate and shrink: \code{\link{rotate}},
  \code{\link{shrink}}
}
\concept{Rotate, translate and shrink}
