% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grow.R
\name{fractal_tree}
\alias{fractal_tree}
\title{Grow a tree}
\usage{
fractal_tree(leaf, depth, growth_fraction)
}
\arguments{
\item{leaf}{leaf shape, a matrix with 4 columns for x, y, xend, yend}

\item{depth}{Integer that specifies the depth (height) of the tree, i.e. the number of loops}

\item{growth_fraction}{}
}
\description{
Grow a tree
}
