% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate.R
\name{translate}
\alias{translate}
\title{Translate object in x and y plane}
\usage{
translate(m, xy)
}
\arguments{
\item{m}{Input matrix}

\item{xy}{a Vector of length two with x and y component}
}
\value{
foo
}
\description{
Translate object in x and y plane
}
\seealso{
Other Rotate, translate and shrink: 
\code{\link{kaleidoscope}()},
\code{\link{rotate}()},
\code{\link{shrink}()}
}
\concept{Rotate, translate and shrink}
